# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Barbossa API
  description: |
    Family music library management API.

    ## Authentication
    All endpoints except `/api/auth/login` require a valid JWT token.
    Pass token in Authorization header: `Bearer <token>`

    ## User Roles
    - **Regular**: Can browse, heart, download, export own library
    - **Admin**: Can manage users, delete from disk, upload to TorrentLeech
  version: 0.1.9
  contact:
    name: Barbossa
  license:
    name: Private
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://barbossa:8080
    description: Docker internal
...
paths:
...
  /api/artists:
    get:
      tags: [artists]
      summary: List all artists (A-Z)
      operationId: listArtists
      security:
        - bearerAuth: []
      parameters:
        - name: letter
          in: query
          schema:
            type: string
            maxLength: 1
          description: Filter by first letter
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: List of artists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistList'
  /api/artists/{id}:
    get:
      tags: [artists]
      summary: Get artist details
      operationId: getArtist
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArtistId'
      responses:
        '200':
          description: Artist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/artists/{id}/albums:
    get:
      tags: [artists]
      summary: Get artist's albums
      operationId: getArtistAlbums
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArtistId'
      responses:
        '200':
          description: List of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumList'

  # ==========================================================================
  # ALBUMS
  # ==========================================================================
  /api/albums:
    get:
      tags: [albums]
      summary: List albums
      operationId: listAlbums
      security:
        - bearerAuth: []
      parameters:
        - name: artist_id
          in: query
          schema:
            type: integer
          description: Optional artist filter
        - name: letter
          in: query
          schema:
            type: string
            maxLength: 1
          description: Filter by artist name initial (A-Z or #)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: List of albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumList'
  /api/albums/{id}:
    get:
      tags: [albums]
      summary: Get album details
      operationId: getAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [albums, admin]
      summary: Delete album from disk (admin only)
      operationId: deleteAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/albums/{id}/tracks:
    get:
      tags: [albums]
      summary: Get album tracks
      operationId: getAlbumTracks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: List of tracks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackList'
...
  /api/search:
    get:
      tags: [artists, albums, tracks]
      summary: Search library
      operationId: searchLibrary
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: type
          in: query
          schema:
            type: string
            enum: [artist, album, track, all]
            default: all
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'

  # ==========================================================================
  # USER LIBRARY
  # ==========================================================================
  /api/me/library:
    get:
      tags: [library]
      summary: Get user's library (hearted albums)
      operationId: getUserLibrary
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: User's hearted albums
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLibrary'
  /api/me/library/albums/{id}:
    post:
      tags: [library]
      summary: Heart album (add to user library)
      operationId: heartAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album hearted, symlinks created
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [library]
      summary: Unheart album (remove from user library)
      operationId: unheartAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album unhearted, symlinks removed
  /api/me/library/tracks/{id}:
    post:
      tags: [library]
      summary: Heart individual track
      operationId: heartTrack
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Track hearted
    delete:
      tags: [library]
      summary: Unheart track
      operationId: unheartTrack
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Track unhearted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ArtistId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    AlbumId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    TrackId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    Forbidden:
      description: Forbidden (admin only)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    Artist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        album_count:
          type: integer
        artwork_path:
          type: string
    ArtistList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
    Album:
      type: object
      properties:
        id:
          type: integer
        artist_id:
          type: integer
        artist_name:
          type: string
        title:
          type: string
        year:
          type: integer
        artwork_path:
          type: string
        total_tracks:
          type: integer
        available_tracks:
          type: integer
        source:
          type: string
        is_hearted:
          type: boolean
    AlbumList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        total:
          type: integer
    Track:
      type: object
      properties:
        id:
          type: integer
        album_id:
          type: integer
        title:
          type: string
        track_number:
          type: integer
        disc_number:
          type: integer
        duration:
          type: integer
        format:
          type: string
        sample_rate:
          type: integer
        bit_depth:
          type: integer
        is_lossy:
          type: boolean
        source:
          type: string
        source_quality:
          type: string
        is_hearted:
          type: boolean
    TrackList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        total:
          type: integer
...
    SearchResults:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: '#/components/schemas/Artist'
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
    UserLibrary:
      type: object
      properties:
        albums:
          type: array
          items:
            $ref: '#/components/schemas/Album'
        total:
          type: integer
...